Struct solana_program_test::ProgramTestContext

pub struct ProgramTestContext {
    pub banks_client: BanksClient,
    pub last_blockhash: Hash,
    pub payer: Keypair,
    pub(crate) genesis_config: GenesisConfig,
    pub(crate) bank_forks: Arc<RwLock<BankForks>>,
    pub(crate) block_commitment_cache: Arc<RwLock<BlockCommitmentCache>>,
    pub(crate) _bank_task: DroppableTask<()>,
}

Fields
banks_client: BanksClientlast_blockhash: Hashpayer: Keypairgenesis_config: GenesisConfigbank_forks: Arc<RwLock<BankForks>>block_commitment_cache: Arc<RwLock<BlockCommitmentCache>>_bank_task: DroppableTask<()>
Implementations

impl ProgramTestContext

pub(crate) fn new( bank_forks: Arc<RwLock<BankForks>>, block_commitment_cache: Arc<RwLock<BlockCommitmentCache>>, banks_client: BanksClient, last_blockhash: Hash, genesis_config_info: GenesisConfigInfo ) -> Self

pub fn genesis_config(&self) -> &GenesisConfig

pub fn increment_vote_account_credits( &mut self, vote_account_address: &Pubkey, number_of_credits: u64 )

Manually increment vote credits for the current epoch in the specified vote account to simulate validator voting activity

pub fn set_account(&mut self, address: &Pubkey, account: &AccountSharedData)

Create or overwrite an account, subverting normal runtime checks.

This method exists to make it easier to set up artificial situations that would be difficult to replicate by sending individual transactions. Beware that it can be used to create states that would not be reachable by sending transactions!

pub fn set_sysvar<T: SysvarId + Sysvar>(&self, sysvar: &T)

Create or overwrite a sysvar, subverting normal runtime checks.

This method exists to make it easier to set up artificial situations that would be difficult to replicate on a new test cluster. Beware that it can be used to create states that would not be reachable under normal conditions!

pub fn warp_to_slot(&mut self, warp_slot: Slot) -> Result<(), ProgramTestError>

Force the working bank ahead to a new slot

pub async fn get_new_latest_blockhash(&mut self) -> Result<Hash>

Get a new latest blockhash, similar in spirit to RpcClient::get_latest_blockhash()