Struct solana_program_test::ProgramTest

pub struct ProgramTest {
    pub(crate) accounts: Vec<(Pubkey, AccountSharedData)>,
    pub(crate) builtins: Vec<Builtin>,
    pub(crate) compute_max_units: Option<u64>,
    pub(crate) prefer_bpf: bool,
    pub(crate) use_bpf_jit: bool,
    pub(crate) deactivate_feature_set: HashSet<Pubkey>,
    pub(crate) transaction_account_lock_limit: Option<usize>,
}

Fields
accounts: Vec<(Pubkey, AccountSharedData)>builtins: Vec<Builtin>compute_max_units: Option<u64>prefer_bpf: booluse_bpf_jit: booldeactivate_feature_set: HashSet<Pubkey>transaction_account_lock_limit: Option<usize>
Implementations

impl ProgramTest

pub fn new( program_name: &str, program_id: Pubkey, process_instruction: Option<ProcessInstructionWithContext> ) -> Self

Create a ProgramTest.

This is a wrapper around default and add_program. See their documentation for more details.

pub fn prefer_bpf(&mut self, prefer_bpf: bool)

Override default SBF program selection

pub fn set_compute_max_units(&mut self, compute_max_units: u64)

Override the default maximum compute units

pub fn set_transaction_account_lock_limit( &mut self, transaction_account_lock_limit: usize )

Override the default transaction account lock limit

pub fn set_bpf_compute_max_units(&mut self, bpf_compute_max_units: u64)

๐Deprecated since 1.8.0: please use set_compute_max_units instead

Override the SBF compute budget

pub fn use_bpf_jit(&mut self, use_bpf_jit: bool)

Execute the SBF program with JIT if true, interpreted if false

pub fn add_account(&mut self, address: Pubkey, account: Account)

Add an account to the test environment

pub fn add_account_with_file_data( &mut self, address: Pubkey, lamports: u64, owner: Pubkey, filename: &str )

Add an account to the test environment with the account data in the provided filename

pub fn add_account_with_base64_data( &mut self, address: Pubkey, lamports: u64, owner: Pubkey, data_base64: &str )

Add an account to the test environment with the account data in the provided as a base 64 string

pub fn add_program( &mut self, program_name: &str, program_id: Pubkey, process_instruction: Option<ProcessInstructionWithContext> )

Add a SBF program to the test environment.

program_name will also be used to locate the SBF shared object in the current or fixtures directory.

If process_instruction is provided, the natively built-program may be used instead of the SBF shared object depending on the BPF_OUT_DIR environment variable.

pub fn add_builtin_program( &mut self, program_name: &str, program_id: Pubkey, process_instruction: ProcessInstructionWithContext )

Add a builtin program to the test environment.

Note that builtin programs are responsible for their own stable_log output.

pub fn deactivate_feature(&mut self, feature_id: Pubkey)

Deactivate a runtime feature.

Note that all features are activated by default.

pub(crate) fn setup_bank( &self ) -> (Arc<RwLock<BankForks>>, Arc<RwLock<BlockCommitmentCache>>, Hash, GenesisConfigInfo)

pub async fn start(self) -> (BanksClient, Keypair, Hash)

pub async fn start_with_context(self) -> ProgramTestContext

Start the test client

Returns a BanksClient interface into the test environment as well as a payer Keypair with SOL for sending transactions